+++ src/ui/static/js/global_config.js
@@ -282,10 +282,58 @@ class Multiple {
               .closest("button")
               .hasAttribute(`data-${this.prefix}-multiple-delete`)
           ) {
+            // We are not removing it really, just hiding it and update values to defaut
+            // By setting default value, group will be send to server and delete (because a setting with default value is useless to keep)
             const multContainer = e.target.closest(
               `[data-${this.prefix}-settings-multiple]`,
             );
-            multContainer.remove();
+            multContainer.classList.add("hidden-multiple");
+            // get setting container
+            const settings = multContainer.querySelectorAll(
+              `[data-setting-container]`,
+            );
+            settings.forEach((setting) => {
+              // for regular input
+              try {
+                const inps = setting.querySelectorAll("input");
+                inps.forEach((inp) => {
+
+                  // case checkbox
+                  if (inp.getAttribute("type") === "checkbox") {
+                    const defaultVal = inp.getAttribute("data-default") || "";
+                    
+                    if (defaultVal === "yes" && !inp.checked) {
+                      inp.click();
+                    }
+
+                  }
+
+                  // case regular
+                  if (inp.getAttribute("type") !== "checkbox") {
+                    const defaultVal = inp.getAttribute("data-default") || "";
+                    inp.setAttribute("value", defaultVal);
+                    inp.value = defaultVal;
+                  }
+
+                });
+              } catch(e) {
+
+              }
+              // for select
+              try {
+                const selects = setting.querySelectorAll("button[data-setting-select]");
+                selects.forEach((select) => {
+                  const defaultVal = select.getAttribute("data-default") || "";
+                  select.querySelector('data-setting-select-text').setAttribute("data-value", defaultVal);
+                  select.querySelector('data-setting-select-text').textContent = defaultVal;
+                  const dropdown = document.querySelector(`[data-setting-select-dropdown="${select.getAttribute('data-setting-select')}"]`);
+                  dropdown.querySelector(`button[value=${defaultVal}]`).click();
+                });
+              }catch(e) 
+              {
+                
+              }
+            });
           }
           //remove last child
         } catch (err) {}
@@ -469,6 +517,15 @@ class Multiple {
       settingCtnr.setAttribute(
         "id",
         settingCtnr.getAttribute("id").replace("_SCHEMA", suffix),
+      );
+    });
+
+    // rename label
+    const labelEls = schemaCtnrClone.querySelectorAll("label");
+    labelEls.forEach((label) => {
+      label.setAttribute(
+        "for",
+        label.getAttribute("for").replace("_SCHEMA", suffix),
       );
     });